Checking for config file before deploying

+++ tutorials/gcp-cos-clamav/start.py
@@ -5,7 +5,9 @@ from os import path
 
 # Init the configs so they can be maintained from outside the container
 def init_conf(conffile):
-    os.system(f"cp /etc/clamav-custom/{conffile} /etc/clamav/{conffile}")
+    if not os.path.exists(f"/etc/clamav/{conffile}"):
+        print(f"Initializing conf file {conffile}")
+        os.system(f"cp /etc/clamav-custom/{conffile} /etc/clamav/{conffile}")
 
 init_conf("clamd.conf")
 init_conf("freshclam.conf")
